skylogix_weather/requirements.txt
requests 
pymongo 
pandas
dotenv
psycopg2-binary
apache-airflow 
sqlalchemy
psycopg2
